+++ packages/main/src/plugin/extension/extension-loader.ts
@@ -31,7 +31,7 @@ import {
 } from '/@/plugin/kubernetes/kube-generator-registry.js';
 import { MenuRegistry } from '/@/plugin/menu-registry.js';
 import { NavigationManager } from '/@/plugin/navigation/navigation-manager.js';
-import { isAsyncFunction } from '/@/plugin/util/deferred.js';
+import { isAsyncFunction } from '/@/plugin/util/async.js';
 import { WebviewRegistry } from '/@/plugin/webview/webview-registry.js';
 import { ApiSenderType } from '/@api/api-sender/api-sender-type.js';
 import { IAsyncDisposable } from '/@api/async-disposable.js';
